readBucket/requirements.txt
# Function dependencies
google-cloud-storage
google-cloud-pubsub
functions-framework==3.*